op-chain-ops: rename from state-surgery
