+++ op-chain-ops/README.md
@@ -1,4 +1,4 @@
-# state-surgery
+# op-chain-ops
 
 This package performs state surgery. It takes the following input:
