+++ op-chain-ops/hardhat/testdata/artifacts/build-info/41b5106372a301360350245ee188494f.json
