+++ op-chain-ops/hardhat/testdata/artifacts/contracts/HelloWorld.sol/HelloWorld.json
