+++ op-chain-ops/hardhat/types.go
@@ -3,7 +3,7 @@ package hardhat
 import (
 	"encoding/json"
 
-	"github.com/ethereum-optimism/optimism/state-surgery/solc"
+	"github.com/ethereum-optimism/optimism/op-chain-ops/solc"
 	"github.com/ethereum/go-ethereum/accounts/abi"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
