+++ op-chain-ops/state/state_test.go
@@ -13,9 +13,9 @@ import (
 	"github.com/ethereum/go-ethereum/core"
 	"github.com/ethereum/go-ethereum/crypto"
 
-	"github.com/ethereum-optimism/optimism/state-surgery/solc"
-	"github.com/ethereum-optimism/optimism/state-surgery/state"
-	"github.com/ethereum-optimism/optimism/state-surgery/state/testdata"
+	"github.com/ethereum-optimism/optimism/op-chain-ops/solc"
+	"github.com/ethereum-optimism/optimism/op-chain-ops/state"
+	"github.com/ethereum-optimism/optimism/op-chain-ops/state/testdata"
 
 	"github.com/stretchr/testify/require"
 )
