--- state-surgery/go.mod
@@ -1,17 +1,0 @@
-module github.com/ethereum-optimism/optimism/state-surgery
-
-go 1.18
-
-require (
-	github.com/ethereum/go-ethereum v1.10.17
-	github.com/mattn/go-isatty v0.0.12
-	github.com/urfave/cli/v2 v2.3.0
-)
-
-require (
-	github.com/cpuguy83/go-md2man/v2 v2.0.0-20190314233015-f79a8a8ca69d // indirect
-	github.com/go-stack/stack v1.8.1 // indirect
-	github.com/russross/blackfriday/v2 v2.0.1 // indirect
-	github.com/shurcooL/sanitized_anchor_name v1.0.0 // indirect
-	golang.org/x/sys v0.0.0-20220503163025-988cb79eb6c6 // indirect
-)
